Move creating item storage representation to update item class CIRC-1418
